+++ hardhat-config/imports.config.js
@@ -1,8 +1,8 @@
 module.exports = {
   imports: [
-    'node_modules/@animoca/ethereum-contracts-assets-1.1.5/artifacts',
+    'node_modules/@animoca/ethereum-contracts-assets-2.0.0/artifacts',
     'node_modules/@animoca/ethereum-contracts-core-1.1.2/artifacts',
     'node_modules/@animoca/ethereum-contracts-sale-2.0.0/artifacts',
-    'node_modules/ethereum-universal-forwarder/export/artifacts',
+    'node_modules/ethereum-universal-forwarder-1.0.0/export/artifacts',
   ],
 };